Share one launch harness across the end-to-end tests
Moves the fixture out of the launch test and into `tests/lib/app.ts`, where
`useApp()` launches the app once for the file that calls it and hands every
test in that file the same window. Every comment that came with the fixture
moves with it; nothing about how the app is launched changes.

One launch per file rather than per test, and not because a launch is
expensive — it takes about a second. It is that a file's tests are steps
through one surface, and reading them as a single session beats relaunching
between each. What that costs is a trace per file instead of per test, which
is the right unit anyway when one launch produced all of them. Anything
needing an app of its own, such as a different seeded config or a restart,
gets a file of its own.

`goto()` and `readConfig()` come along because the tests that follow all
navigate by hash and then read what the main process wrote. `goto()` rebuilds
the URL rather than replacing it, since the query string carries the accounts
handed over at load and dropping it would reload the app without them.
